kelly_audio2face/requirements.txt
# Kelly Audio2Face-3D Requirements
# Core dependencies for Kelly avatar facial animation

# Audio2Face-3D dependencies
numpy==1.26.4
scipy==1.13.0
grpcio==1.72.0rc1
protobuf==4.24.1
PyYAML==6.0.1
pandas==2.2.2

# Additional Kelly-specific dependencies
pathlib2==2.3.7
asyncio-mqtt==0.16.1
aiofiles==23.2.1
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
black==23.9.1
flake8==6.1.0
